Reorganizando rotas e home para agendamento
