+++ src/app/component/agendamento/agendamento.component.html
@@ -1,0 +1,1 @@
+<router-outlet></router-outlet>
