+++ src/app/component/agendamento/agendamento.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'agendamento',
+  templateUrl: './agendamento.component.html'
+})
+
+export class AgendamentoComponent {
+  constructor() { }
+}
