--- src/app/component/agendamento/home/agendamento.component.ts
@@ -1,16 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'agendamento',
-  templateUrl: './agendamento.component.html',
-  styleUrls: ['./agendamento.component.css']
-})
-
-export class AgendamentoComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit(): void {
-    
-  }
-}
