+++ src/app/component/agendamento/home/home-agendamento.component.html
@@ -1,0 +1,67 @@
+<mat-tab-group>
+    <mat-tab label="Agendamentos confirmados">
+        <mat-table [dataSource]="agendamentosConfirmados" class="mat-elevation-z8">
+            
+            <ng-container matColumnDef="imovel">
+              <mat-header-cell *matHeaderCellDef> Imóvel </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.imovel}} </mat-cell>
+            </ng-container>
+          
+            <ng-container matColumnDef="date">
+              <mat-header-cell *matHeaderCellDef> Data Agendada </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.date}} </mat-cell>
+            </ng-container>
+          
+            <ng-container matColumnDef="keyHourStart">
+              <mat-header-cell *matHeaderCellDef> Chave Disponível </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.keyHourStart}} </mat-cell>
+            </ng-container>
+
+            <ng-container matColumnDef="keyHourEnd">
+              <mat-header-cell *matHeaderCellDef> Devolução da Chave </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.keyHourEnd}} </mat-cell>
+            </ng-container>
+
+            <ng-container matColumnDef="Status">
+              <mat-header-cell *matHeaderCellDef> Status </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.status}} </mat-cell>
+            </ng-container>
+          
+            <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
+            <mat-row *matRowDef="let row; columns: displayedColumns;"></mat-row>
+          </mat-table>
+    </mat-tab>
+    
+    <mat-tab label="Agendamentos pendentes">
+        <mat-table [dataSource]="agendamentosPendentes" class="mat-elevation-z8">
+            
+            <ng-container matColumnDef="imovel">
+              <mat-header-cell *matHeaderCellDef> Imóvel </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.imovel}} </mat-cell>
+            </ng-container>
+          
+            <ng-container matColumnDef="date">
+              <mat-header-cell *matHeaderCellDef> Data Agendada </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.date}} </mat-cell>
+            </ng-container>
+          
+            <ng-container matColumnDef="keyHourStart">
+              <mat-header-cell *matHeaderCellDef> Chave Disponível </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.keyHourStart}} </mat-cell>
+            </ng-container>
+
+            <ng-container matColumnDef="keyHourEnd">
+              <mat-header-cell *matHeaderCellDef> Devolução da Chave </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.keyHourEnd}} </mat-cell>
+            </ng-container>
+
+            <ng-container matColumnDef="Status">
+              <mat-header-cell *matHeaderCellDef> Status </mat-header-cell>
+              <mat-cell *matCellDef="let element"> {{element.status}} </mat-cell>
+            </ng-container>
+          
+            <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
+            <mat-row *matRowDef="let row; columns: displayedColumns;"></mat-row>
+          </mat-table>
+    </mat-tab>
+</mat-tab-group>
