+++ src/app/component/agendamento/home/home-agendamento.component.ts
@@ -1,0 +1,34 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'home-agendamento',
+  templateUrl: './home-agendamento.component.html',
+  styleUrls: ['./home-agendamento.component.css']
+})
+
+export class HomeAgendamentoComponent implements OnInit {
+  agendamentosConfirmados = [
+    {
+      imovel: '',
+      date: '',
+      keyHourStart: '',
+      keyHourEnd: '',
+      Status: ''
+    }
+  ];
+  agendamentosPendentes = [
+    {
+      imovel: '',
+      date: '',
+      keyHourStart: '',
+      keyHourEnd: '',
+      Status: ''
+    }
+  ];
+  displayedColumns: string[] = ['imovel', 'date', 'keyHourStart', 'keyHourEnd', 'Status'];
+  constructor() { }
+
+  ngOnInit(): void {
+
+  }
+}
