+++ src/app/component/agendamento/novo-agendamento/novo-agendamento.component.css
@@ -1,0 +1,8 @@
+.container {
+    display: flex;
+    justify-content: center;
+}
+
+form {
+    margin-top: 1.2rem;
+}
